requirements.txt
streamlit==1.34.0
pillow==10.3.0
pytube==15.0.0
